# cache/routes/api/v1/cache/[tenant]/narinfo.ts
/*
 * Response format:
 * Array<string>
 *
 * Return format:
 * Array<string>
 * */

import { Derivation_tenant_link, Requests, Tenants } from "@iglu-sh/shared/db";
import { Logger } from "@iglu-sh/shared/logger";
import type { openapi_definiton, request, tenant } from "@iglu-sh/shared/types";
import { Authentication, IPFiltering, MakeRestResponse } from "@iglu-sh/shared/utils";
import type { Request, Response } from "express";
import bodyParser from "express";
import { z } from "zod";
import { error_response_schema } from "@/shared/utils/zod/zod_rest_schemas";

const body_format = z.array(z.string());

export const openapi: openapi_definiton = {
    meta: {
        path: "/api/v1/cache/{tenant}/narinfo",
        authentication_required: true,
        tags: ["api/v1/cache", "cachix"],
    },
    routes: [
        {
            method: "post",
            description: "Checks if a given array of narhashes is already cached",
            summary: "Checks if a given array of narhashes is already cached",
            request: {
                params: z.object({
                    tenant: z.string(),
                }),
                body: {
                    description:
                        "Array of Nix Store Hashes that should be checked if they are already cached",
                    required: true,
                    content: {
                        "application/json": {
                            schema: body_format,
                        },
                    },
                },
            },
            responses: {
                200: {
                    description: "A list of hashes that are not already stored in the tenant",
                    content: {
                        "application/json": {
                            schema: z.array(z.string()),
                        },
                    },
                },
                400: {
                    description:
                        "Invalid Body: This is returned when your request body is not in the proper format",
                    content: {
                        "application/json": {
                            schema: error_response_schema,
                        },
                    },
                },
                404: {
                    description: "Returned when the requested tenant does not exist on the server",
                    content: {
                        "application/json": {
                            schema: error_response_schema,
                        },
                    },
                },
            },
        },
    ],
};

export const post = [
    IPFiltering(),
    Authentication(),
    bodyParser.json({ limit: "50mb" }),
    async (req: Request, res: Response) => {
        const TENANT_NAME = req.params.tenant;
        if (!TENANT_NAME || Array.isArray(TENANT_NAME)) {
            Logger.debug("Narinfo request did not contain tenant");
            return res.status(404).json(
                MakeRestResponse(404, "Not found", true, {
                    error_details: "Did not find cache with the given name",
                }),
            );
        }

        const BODY_ARRAY = req.body;
        if (!BODY_ARRAY || !Array.isArray(BODY_ARRAY)) {
            return res.status(400).json(
                MakeRestResponse(400, "Invalid Body", true, {
                    error_details: "The Request body is malformed",
                }),
            );
        }

        const verified_body_array = body_format.safeParse(BODY_ARRAY);
        if (!verified_body_array.success) {
            return res.status(400).json(
                MakeRestResponse(400, "Invalid Body", true, {
                    error_details: "The Request body is malformed",
                }),
            );
        }

        // Get the tenant from the database
        const tenant = await new Tenants().getByName(TENANT_NAME).then((res) => {
            return res[0] as tenant;
        });

        // Get the stored records from the array in the body
        const hashes_stored_in_db = await new Derivation_tenant_link().getByNixStoreHashes(
            verified_body_array.data,
            tenant.id,
        );

        if (hashes_stored_in_db.length === 0) {
            return res.status(200).json(verified_body_array.data);
        }
        const hashes_as_string: Array<string> = [];
        const requests_to_insert: Array<request> = [];
        for (const link of hashes_stored_in_db) {
            requests_to_insert.push({
                id: "n/a",
                derivations_tenants_links: link.id,
                date: Date.now(),
                direction: "inbound",
                url: `/api/v1/cache/${link.tenants_id.name}/narinfo`,
            });
            hashes_as_string.push(link.derivations_id.cstorehash);
        }
        await new Requests().bulk_insert(requests_to_insert);
        const hashes_not_in_db: Array<string> = verified_body_array.data.filter((x) => {
            return !hashes_as_string.includes(x);
        });

        // Get the hashes not in the database
        return res.status(200).json(hashes_not_in_db);
    },
];