# openapi/schema.openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "0.0.1",
        "title": "Iglu API",
        "description": "Iglu API Docs for use in your development"
    },
    "servers": [{ "url": "/" }],
...
    "paths": {
...
        "/api/v1/cache/{tenant}/narinfo": {
            "post": {
                "description": "Checks if a given array of narhashes is already cached",
                "summary": "Checks if a given array of narhashes is already cached",
                "tags": ["api/v1/cache", "cachix"],
                "parameters": [
                    {
                        "schema": { "type": "string" },
                        "required": true,
                        "name": "tenant",
                        "in": "path"
                    },
                    {
                        "schema": { "type": "string" },
                        "required": true,
                        "name": "authorization",
                        "in": "header"
                    },
                    {
                        "schema": { "type": "string" },
                        "required": true,
                        "name": "user-agent",
                        "in": "header"
                    }
                ],
                "requestBody": {
                    "description": "Array of Nix Store Hashes that should be checked if they are already cached",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": { "type": "array", "items": { "type": "string" } }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "A list of hashes that are not already stored in the tenant",
                        "content": {
                            "application/json": {
                                "schema": { "type": "array", "items": { "type": "string" } }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Body: This is returned when your request body is not in the proper format",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status_code": { "type": "number" },
                                        "status_message": { "type": "string" },
                                        "is_error": { "type": "boolean" },
                                        "data": {
                                            "type": "object",
                                            "properties": { "error_details": { "type": "string" } },
                                            "required": ["error_details"]
                                        },
                                        "timestamp": { "type": "number" }
                                    },
                                    "required": [
                                        "status_code",
                                        "status_message",
                                        "is_error",
                                        "data",
                                        "timestamp"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "The key you were using is not recognized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status_code": { "type": "number" },
                                        "status_message": { "type": "string" },
                                        "is_error": { "type": "boolean" },
                                        "data": {
                                            "type": "object",
                                            "properties": { "error_details": { "type": "string" } },
                                            "required": ["error_details"]
                                        },
                                        "timestamp": { "type": "number" }
                                    },
                                    "required": [
                                        "status_code",
                                        "status_message",
                                        "is_error",
                                        "data",
                                        "timestamp"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - No authorization, either you didn't provide an authorization header or your header is not in the Bearer format",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status_code": { "type": "number" },
                                        "status_message": { "type": "string" },
                                        "is_error": { "type": "boolean" },
                                        "data": {
                                            "type": "object",
                                            "properties": { "error_details": { "type": "string" } },
                                            "required": ["error_details"]
                                        },
                                        "timestamp": { "type": "number" }
                                    },
                                    "required": [
                                        "status_code",
                                        "status_message",
                                        "is_error",
                                        "data",
                                        "timestamp"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The requested tenant does not exist on this server",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status_code": { "type": "number" },
                                        "status_message": { "type": "string" },
                                        "is_error": { "type": "boolean" },
                                        "data": {
                                            "type": "object",
                                            "properties": { "error_details": { "type": "string" } },
                                            "required": ["error_details"]
                                        },
                                        "timestamp": { "type": "number" }
                                    },
                                    "required": [
                                        "status_code",
                                        "status_message",
                                        "is_error",
                                        "data",
                                        "timestamp"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}